+++ example/source/components/UI/Form.jsx
@@ -1,3 +1,3 @@
-import { createForm } from 'react-form-validation';
+import { createForm } from '@feerzlay/react-form-validation';
 
 export default createForm();