+++ webpack.config.js
@@ -1,0 +1,42 @@
+require('dotenv').config();
+
+const path = require('path');
+const webpack = require('webpack');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+
+module.exports = {
+  mode: 'production',
+  devtool: 'cheap-eval-source-map',
+  entry: './source/index.js',
+  output: {
+    filename: 'index.js',
+    path: path.resolve(__dirname, 'out'),
+    libraryTarget: 'commonjs2'
+  },
+  optimization: {
+    minimizer: [new UglifyJsPlugin()]
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        loader: 'babel-loader',
+        exclude: [/node_modules/]
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['.js', '.jsx']
+  },
+  externals: {
+    react: 'commonjs react',
+    'prop-types': 'commonjs prop-types'
+  },
+  plugins: [
+    new webpack.NoEmitOnErrorsPlugin(),
+    new BundleAnalyzerPlugin({
+      analyzerMode: +process.env.OPEN_BUNDLE_ANALYZER ? 'server' : 'disabled'
+    })
+  ]
+};