+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.4.0] - 2018-09-16
 - Add validation event config
 - Use comma to separate multiple rule agruments
 - Add addError and removeError public methods